+++ examples/basic/src/my-javascript-component-no-props/my-javascript-component-no-props.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MyJavascriptComponentNoProps = () => {
+    return <div>The name is test!</div>;
+};
+
+export default MyJavascriptComponentNoProps;